Add readonly opcodes
Summary:
This adds the readonly opcodes but does not implement
any related functionality. The existence of the opcodes allows
us to enable some foundational bytecode analysis functionality needed
to turn on the JIT later in the stack. No compilers should
be emitting the opcodes yet, so we should be safe to define them without
an implementation.

I didn't run `make regen-opcode-targets` since there are no implementations
of the opcodes yet.

Reviewed By: itamaro

Differential Revision: D37968990

fbshipit-source-id: dd3a51fcc4270e11426197af22a21fcea113b420