Remove JIT support for `{BEGIN, END, CALL, POP}_FINALLY`
Summary:
These were removed in 3.9.

We'll need to take a closer look at handling `try/finally` in 3.10
but it looks like Mark Shannon implemented the JIT's strategy for
handling `try/finally` (duplicating code in the finally block) in
the bytecode compiler in 3.9 (https://bugs.python.org/issue33387).
That should make things a lot easier for us when we get to supporting
`try/finally`.

Reviewed By: itamaro

Differential Revision: D37968997

fbshipit-source-id: 216d756e272b7d62df5deac5d9641fb0b60271d9