Remove JIT support for `WITH_CLEANUP_{START,FINISH}`
Summary:
These were removed in 3.9.

These were removed in 3.9 by Mark Shannon when he got rid of the
various `*_FINALLY` opcodes in https://bugs.python.org/issue33387.
`with` statements are now implemented using `RERAISE` and `WITH_EXCEPT_START`.
Like the `*_FINALLY` opcodes, we'll need to take a closer look
at compiling code with `with` statements, but I expect the implementation
to be simpler than what we had for 3.8.

Reviewed By: itamaro

Differential Revision: D37968994

fbshipit-source-id: d80bbd684d202441ab0830a6d0c4147637c3eeac