Add a stub definition of `PyArray_Type`
Summary:
In order to get the basic end-to-end compilation/execution flow for
the JIT working we need to either define `PyArray_Type` or remove
the places in the JIT that reference it. I've chosen to define it.
It's referenced in quite a few places (the type system and optimization passes),
and we ultimately want the functionality provided by the referrring
code. If we were to remove the references to get the JIT building we'd
end up having to add them back in.

For now I've just added a stub definition. The actual definition in
3.8 was provided by moving the `array` module from `Modules/` into
core CPython, however, in 3.10 `75bf107c` moves the `PyArray_Type`
into module state that is initialized dynamically and stored on
the module object. We can decide how to expose the real `PyArray_Type`
to the JIT when we get to porting Static Python features.

Reviewed By: itamaro

Differential Revision: D37968992

fbshipit-source-id: a3a489e5e21516c538b843902ab2f37e255daf48