Get a minimal end-to-end JIT compilation flow working
Summary:
This is a grab bag of the remaining changes that were required to get
the end-to-end JIT compilation flow working. We only allow `LOAD_CONST`,
`LOAD_FAST`, `RETURN_VALUE`, and `STORE_FAST` right now, but that's enough
to compile and execute trivial functions. We should be able to expand the
set quite quickly.

Things of note:

1. I changed the stub for `Py_IS_IMMORTAL` to always return `false`. This
   should be the correct behavior until we implement immortal references.
2. I had to restore an `extern "C"` guard to `Include/cpython/pystate.h`
   that was removed.

From here we should start focusing on enabling more opcodes and ensuring
that other features of the JIT are working (e.g. deopt, shadow frame mode,
the inliner).

I may look at porting `PyCodeObject::co_qualname` because it's needed by
the JIT list, and we need thte JIT list in order to use `jitlist_bisect.py`,
which is super helpful for minimizing repros of JIT issues.

Reviewed By: tekknolagi

Differential Revision: D37968991

fbshipit-source-id: 02042bec295654b461d3fbe508b2f6dc1ba8532b