Port `Add qualname to codeobject`
Summary:
This brings over the change adding the `co_qualname` field to code
objects. This was mostly straightforward, the only difference worth
calling out from the original change is that I had to modify the
regex used in one of the tests in `test_compiler` that verifies
recursion limits are enforced. With this change it looks like the
recursion limit is hit in a slightly different place and the error
message is a little different.

I also had to bring over the `cinder` module in order to bring over
the test from the original diff. The qualname change should be easy
to extract on its own for upstreaming so I've tagged this as such.

Reviewed By: carljm

Differential Revision: D38003337

fbshipit-source-id: ec735270190a9204760383d912de6a7024704383