Get `test_cinderjit` passing when run under the JIT
Summary:
Many functions in `test_cinderjit` are decorated with `failUnlessJITCompiled`,
which raises an exception if the JIT is enabled and the function cannot be
compiled. Additions of the decorator were predicated on the assumption that
the JIT should be able to compile the decorated function. As we gradually
expand the set of supported opcodes in the JIT that assumption may no longer
hold. This diff modifies the decorator to only ensure that we can compile
the decorated function if we support all its opcodes. This should ensure
that we're automatically enforcing JIT compilation of functions as we
expand opcode support.

Reviewed By: tekknolagi

Differential Revision: D38065381

fbshipit-source-id: 8b93cbd0db2dd81752075fe2a13a49f9f9e895e7