Enable the JUMP family of opcodes in the JIT
Summary:
The only thing to note, though a major one, is that I discovered
that in 3.9/3.10 CPython started using number of instructions
instead of byte offsets to specify the value of the interpreter's
instruction pointer both in the bytecode and in `PyFrameObject`
(i.e. the units for f_lasti are now instructions, not bytes).
Practically, this means that everything is divided by two.

The JIT currently uses byte offsets everywhere (this matches
the behavior of 3.8). For now I've kept things the same and
performed the translation at the places where the JIT has to
interface with the rest of CPython (deopt). At some point we
should go through and update the JIT to match CPython.

Reviewed By: jbower-fb

Differential Revision: D38087708

fbshipit-source-id: e543a6fe8d200a382506f0b951660e5f1000a419