Enable stack manipulation opcodes
Summary:
Enable the stack manipulation opcodes (e.g. `POP_TOP`).

The only thing of note is that this enables us to compile simple
loops, which needed support for loading the eval breaker. The
implementation has changed between 3.8 and 3.10. Thankfully, it's
a little simpler for us.

I'm tagging this as `cinder-310-upstream-challenge` but I actually think
that with our own interpreter loop this won't be an issue. The only
function called by the JIT now is `eval_frame_handle_pending`, which is
defined in `ceval.c`. The implementation of will hopefully come over
with the interpreter loop.

Reviewed By: jbower-fb

Differential Revision: D38126314

fbshipit-source-id: 7192ff782e823efca6fc688eb86172a565ce6e2d