Add decorator to assert a function is jitted only if broken tests are enabled
Summary:
D38065381 modified the `failUnlessJITCompiled` decorator to only raise an error
if we supported all of the function's opcodes in the JIT. This provides a nice
way to get `test_cinderjit` running under the JIT and incrementally enforce JIT
compilation as we add support for more opcodes during porting.

However, it introduces a surprising footgun when adding new tests during porting.
If you add a test method that you expect to be JIT-compiled that accidentally
contains a bytecode that is unsupported by the JIT, the decorator will happily
(and silently) accept it.

To remedy this, I've introduced `failUnlessJITCompiledIfBrokenTestsAreEnabled`,
which only enforces JIT compilation if broken tests are enabled. I've modified
all pre-existing uses of `failUnlessJITCompiled` to use this decorator instead.
We can codemod it back and delete the decorator once porting is done.

The `failUnlessJITCompiled` decorator now does what you expect - it always asserts
that the decorated function is JIT-compiled. All new tests that are added during
porting should use this.

Reviewed By: tekknolagi

Differential Revision: D38166836

fbshipit-source-id: fac2e204013dd06df201dfc2f9a56c766a1a4cc3