pretty-print co_lines for side-by-side bytecode compare
Summary:
Minor quality-of-life improvement to `dis_stable` for bytecode comparison.

If `co_lines` is long, pretty-print it with one tuple per line to make it easier to see where
the discrepancy is.

Also don't print `co_linetable` at all, it's just noise, and if it is different then `co_lines`
will be different too.

Reviewed By: itamaro

Differential Revision: D38252963

fbshipit-source-id: 6c9921e758d8bf6409991d9aeb6d1fe2ca140509