fix non-deterministic bug in frozenset const handling
Summary:
Our handling of constant sets had a non-deterministic bug due to dependence on set
iteration order. If two constant sets shared the same elements, we might or might not
(depending on the random hash seed value) share the same entry in `co_consts` for
both of them, or add duplicate entries. Upstream will always share. The fix is to use a
frozenset of the const keys of the elements, rather than a tuple, as part of the const
key for a frozenset, so that it compares order-independent.

Also make `co_consts` discrepancies a bit easier to debug with large `co_consts`
values by pretty-printing it (so if it is large, it goes one-per-line instead of a very long
single line.)

Reviewed By: itamaro

Differential Revision: D38264054

fbshipit-source-id: 9d2aa8e104e95fd99891d01fd89283c50c5847fd