Factor inputs to CFrame.use_tracing into a helper function
Summary:
The `ts->c_tracefunc != NULL || ts->c_profilefunc != NULL` check is
duplicated in a number of places. Pull the check out into a helper function and
call that instead, centralizing the logic.

There were a few places that set `ts->c_tracefunc = NULL` and then only checked
`ts->c_profilefunc` (or vise-versa); replacing those with a call to this new
helper should be equivalent (and I expect the compiler is smart enough to
generate the same code).

Reviewed By: tekknolagi

Differential Revision: D38226936

fbshipit-source-id: 0e8f42b65057617e5d707265cf4272a3dbf99e73