Port Interpreter type profiling
Summary:
This is a port of the upstream parts of D31065920 (https://github.com/facebookincubator/cinder/commit/cf99db9f18e2fdee799643de73d275da592738c1), with a few things
in different places due to changes in upstream 3.10. I also went through new
opcodes and added them to the list of profiled opcodes as appropriate.

I spent some time thinking about ways to minimize the changes to CPython data
structures (`_ts` and `_ceval_state` in this case) and couldn't come up with
anything that felt great. I've left TODO comments for now, and when we get
serious about CinderVM work in the future we can take another look.

One change that we've already run into elsewhere is that
`PyFrameObject::f_lasti` is now a bytecode index, not offset. I corrected for
this in one place in `_PyJIT_ProfileCurrentInstr()`, and kept using offsets in
the rest of the code. As long as we're clear with our terminology and variable
names, this should be fine going forward (and `PyCode_Addr2Line()` still takes
an offset, not an index).

Reviewed By: mpage

Differential Revision: D38227265

fbshipit-source-id: 5c0374a3d5e9a7f4b80262e7c76866a4bca558c7