don't visit stringified annotations in build symtables
Summary:
If the symbol table builder visits annotations when `__future__.annotations` is active,
it will add names to the symbol table that don't belong there, because they aren't
actually referenced at runtime (they are just part of a stringified annotation). In nested
functions or classes with annotations, this can result in a difference in the bytecode,
because we may emit closures for variables that aren't actually used.

Upstream symbol visitor seems to still visit stringified annotations in a special
`AnnotationBlock`, but I'm not sure why -- there can't be anything in a string that
the symbol table would care about. Just skipping the visit as I do here seems to pass
all the tests I can come up with. Maybe later we'll discover a reason why we need
something more.

See e.g.
https://github.com/facebookincubator/cinder/blob/cinder/3.10/Python/symtable.c#L1826-L1838

Reviewed By: itamaro

Differential Revision: D38284939

fbshipit-source-id: c174f623697e1830a186346d9142dabdb9b70337