Improve the cached_property C accelerator
Summary:
The pure-python implementation is a hot function, we should use
the C accelerator if we can. Before we can do so, we need to bring the C
implementation closer to the Python one.

Changes in this diff:
- Updated the cached_property design to work with `__set_name__` (PEP487)
- Introduced some error checks that are present in the Python version
- Refactored the functools tests to also test the C implementation.
(note that a few tests are skipped for the C implementation, since it
actually supports slots, and doesn't support threading).

Reviewed By: carljm

Differential Revision: D37571673

fbshipit-source-id: 452e9c9b26cb0078e9354f6f0d80d42c375792c9