Unwatch types when ICs are destroyed
Summary:
Something about the finalization order changed enough between 3.8 and 3.10 that
it exposed a bug in JIT finalization. We weren't unsubscribing to
changes to types when ICs were destroyed (or clearing the set of subscriptions
during finalization). As a result, we ended up in a scenario where:

1. The JIT is finalized, ICs are destroyed, and memory allocated for them is freed.
2. A type that was watched by an IC is modified.
3. The type watching machinery calls `notifyTypeChanged` with a pointer to an IC
   that was freed in (1). `notifyTypeChanged` ends up writing to memory that was
   freed.

This diff fixes the error by unsubscribing from changes to watched types when ICs
are destroyed. We should also move the global state that contains the set of
watched types into `Runtime`, but that will come after porting is complete in
order to minimize the divergence between 3.8 and 3.10.

Reviewed By: tekknolagi

Differential Revision: D38361246

fbshipit-source-id: 4a76afd002360af1e46967da8900d67a55247166