Set `tstate->frame` to calling frame before deopting
Summary:
On entry, the interpreter loop expects that `tstate->frame` points to the frame of the calling
function, and that it will manage updating `tstate->frame` to point to the supplied frame
while it's executing and reset `tstate->frame` to point to the calling frame when execution
completes. We weren't respecting this contract when deopting. We were leaving `tstate->frame`
pointing to the frame of the deopting function, rather than its caller, but were relying
on the intepreter loop to update `tstate->frame` to point to its caller when done. This is
mostly a non-issue in practice. The interpreter loop performs a recursion check, sets
`tstate->frame` to point to the supplied frame, and then updates `tstate->frame` to
point to the calling frame before returning. However, if the recursion check fails, the
interpreter loop returns early, and won't update `tstate->frame` to point to the calling
frame.

Reviewed By: tekknolagi

Differential Revision: D38361248

fbshipit-source-id: 2009f83f47f5ff7248330a06e33ecc92013a51f1