Pop TOS after slot-with-default is assigned
Summary:
Currently, we don't pop the return value from TOS after a
slot-with-default is assigned to. Because of this, the attached test
case fails at the assertion within `PyFlowGraph().push_block()`.

With this change, we clear the returned value after the method call, no
longer leaving the stack with the extra return value.

Reviewed By: sinancepel

Differential Revision: D38379582

fbshipit-source-id: 3da80c963508d94765380a02df541564c2560bbc