Fix shutdown deadlock in test_current_{exceptions,frames}
Summary:
This is a port of D26111438, also applied to the new test,
`test_current_exceptions`, that fits the same pattern.

These tests spawn a thread that waits on a `threading.Event`. If the tests fail
any of their assertions, the `Event` won't be signaled and the thread will wait
indefinitely, causing a deadlock when `threading._shutdown()` tries to join all
outstanding threads.

Reviewed By: czardoz

Differential Revision: D38362867

fbshipit-source-id: c14869e996d79c99644670f579d1242b8287b446