Support closure-related opcodes in the JIT
Summary:
Most of the test failures were tests we had disabled in 3.8 for
various reasons that need to be disabled in 3.10. I put the disabled tests in a
file on the side rather than modifying the upstream files this time, in
preparation for our CinderVM work.

There are two small code fixes:
- Clear some JIT maps in `_PyJIT_Finalize()` due to some timing differences
  during process shutdown in 3.10. We shouldn't be holding on to strong
  references to Python objects after JIT shutdown anyway.
- Call `set_attribute_error_context()` in `raise_attribute_error`, to match
  upstream changes to `PyObject_GetAttr()`. This also required some changes in
  `ASMGeneratorTest.LoadAttrRaisesError`, since `set_attribute_error_context()`
  calls `PyErr_NormalizeException()`, which reifies a full `AttributeError`
  object.

Reviewed By: tekknolagi

Differential Revision: D38336257

fbshipit-source-id: f4d76b8ba12b18371675ff436ba2001c9c1b7cc6