actually statically compile static dataclass generated methods
Summary:
We custom-generate flow graphs for static dataclass generated methods. The initial implementation missed that these functions needed to be marked as statically compiled. Fix that, since these are indeed static functions we invoke against. As a corollary, we also need to fix the return type descr to appear in the flow graph as an extra const for the classloader/patching infrastructure to work, so let's add that as an API for the flow_graph construction so that it's impossible to miss :)

With this change, static dataclasses should work again in patchable cases, so re-expose static dataclasses to the world.

Reviewed By: carljm

Differential Revision: D38388559

fbshipit-source-id: b3edbd6c249c03bbce2be2b5a7693e5b6dea9c06