Add script for chronos to run fuzzer
Summary:
Added a script `make run_fuzzer_nightly` to be run by chronos nightly. The script runs the fuzzer on a predetermined set of test files `TEST_FILE_LIST`, which currently holds 10 test files. `TEST_FILE_LIST` is subject to change. Currently, the script is set to spawn 4 subprocesses for each function, but this is also subject to change.

[This chronos job](https://www.internalfb.com/intern/chronos/job/?jobname=cinder-run-fuzzer-nightly&smc=chronos_gp_admin_client), which runs `make run_fuzzer_nightly` once every day at midnight, will be enabled if / when this diff gets merged into master.

Reviewed By: sinancepel

Differential Revision: D38324730

fbshipit-source-id: 6743d2952e6a2ce5cb83088d125455caf1b519fe