Add Python test features making it easier to enable tests when features are ported
Summary:
Adds a giant enum with features which are not yet ported (mostly JIT opcodes). This can be used with new decorators to skip tests or skip JIT for tests which require these features. Whenever a feature is ported it can be removed from the enum and it will now be easy to find tests which should now work.

These new decorators replace `failUnlessJITCompiledIfBrokenTestsEnabled()`.

The real "magic" in this diff was adding a hacky generator `dumpMissingOpcodes()` to determine missing opcodes in functions which are supposed to be JIT compiled. I processed the output of this with another hacky script to automatically annotate `test_cinderjit.py`: P519967686. The rest of the diff was mostly just manual hard work.

Note I've only addressed `test_cinderjit.py` in this diff although we can probably also use this to replace some instance of `cinderPortingBrokenTest()` in other places.

Reviewed By: mpage

Differential Revision: D38373257

fbshipit-source-id: 3ce135e4ec14c470c4fb97927cd11206fa1a54be