Add `BCOffset` and `BCIndex` types to avoid unit mismatch
Summary:
Bytecode offsets are much lower stakes than flying probes (in)to Mars,
but we can still do better than JPL and Lockheed Martin did in 1998. The JIT
uses byte offsets and instruction indexes in a few different places, so this
diff adds unit-aware types to take the place of plain ints.

The new types support common arithmetic operations and implicit conversion
between each other, but don't allow implicit conversion to or from raw integral
values. This is motivated by CPython 3.10 switching from byte offsets to
instruction indexes in many places, increasing the odds of making a subtle
mistake.

I decided to do this as 3.10-only because it's good to have it on 3.10 sooner
rather than later, and doing it on 3.8 head then porting to 3.10 could create a
variety of conflicts due to our 3.10 fork point being a couple months old. This
may cause conflicts as we port newer 3.8 diffs after phase 1 of the port is
complete, but forcing us to deal with those conflicts in 3.10 will ensure we
think about whether the relevant code is handling offsets or indexes.

Reviewed By: jbower-fb

Differential Revision: D38394275

fbshipit-source-id: ff90c0ab7830efb4fd445aa119448cd7cae95a1d