Enable exception handling and context manager opcodes in the JIT
Summary:
This unlocks `try/except`, `try/finally`, `raise`, and `with` in the JIT.

The only interesting thing here is I had to fix a couple more instances of
instruction index/offset mismatches.

Mark Shannon did us a big favor and adopted the JIT's approach of duplicating
finally blocks in the CPython compiler. I already deleted what was in the JIT
as part of the initial bring up, so there's not much for us to do!

Reviewed By: tekknolagi

Differential Revision: D38372826

fbshipit-source-id: 52b0f435eaba35f8c304c9a1e06c2649fb6b682d