fix lineno in strict rewriter and codegen
Summary:
Make sure the strict rewriter provides line numbers for newly generated AST nodes.
For most places the line number is -1 indicating that the generated code is following the previous line number.
In function and module start, the line numbers are explicitly set to follow the first non-generated node in the body

Reviewed By: sinancepel

Differential Revision: D38413798

fbshipit-source-id: 05c12b2ce1c2fd13989800706137baaf69293490