Support DICT_MERGE, DICT_UPDATE and container opcodes
Summary:
`DICT_MERGE` and `DICT_UPDATE` are used to implement `f(**a, **b)` and
`{**a, **b}`, replacing `BUILD_MAP_UNPACK_WITH_CALL` and `BUILD_MAP_UNPACK`
from 3.8. I tacked the grab-bag of container opcodes on because some of them
were needed to enable any interesting tests of the dict opcodes.

Reviewed By: tekknolagi

Differential Revision: D38448588

fbshipit-source-id: 2faabfda342cff9d5d470e90518c355a339b56a5