skip test that depends on classloader
Summary:
This odd no-op mention of the skip decorator looks like it was
meant to skip the test below. All similar tests of the patching
proxy are skipped, I think because patching a strict module now
calls into classloader to inform static python of the patch. And this
test currently causes a segfault on shutdown (we try to dealloc
`True`.) Skip the test for now until classloader and SP are working.

Differential Revision: D38563393

fbshipit-source-id: 0bace8d3d3bd363cee08bfb0ebfc9426372c185f