make classloader and xxclassloader compile
Summary: Bringing up the classloader and xxclassloader compile is step 0 for being able to properly run the static compiler tests. For now, we disable all the coroutine functionality and do not yet implement proper slot fusing (D26769730). These will be addressed in future diffs as we get more of the static python tests online.

Reviewed By: tekknolagi

Differential Revision: D38476178

fbshipit-source-id: 4c2937baea213ae3ad5b1cf882972e69349422b4