Fix bugs in run_fuzzer_on_test_files
Summary: Fixed bug where file directories were incorrect in `run_fuzzer_on_test_files`, causing it to run the fuzzer on the same file repeatedly. Modified run directory in for `make run_fuzzer_on_test_files` to be in the cinder folder as opposed to the fuzzer folder. Additionally, stdout and stderr output was not appearing on Skycastle since they are stored in `fuzzer_output.txt`, so added a print statement to print file contents to stdout.

Reviewed By: sinancepel

Differential Revision: D38430776

fbshipit-source-id: 4796d748d7ade7961002c9be2f38269f8355602a