add PRIMITIVE_BOX support
Summary: One interesting thing about this diff is that the current Cinder 3.8 implementation that I copied over contains two extra shadow opcodes. I'm not sure if it is bad that I'm introducing these before the next phase, comments welcome as to whether I should purposefully port an older version and re-apply the patches on top later.

Reviewed By: tekknolagi

Differential Revision: D38619154

fbshipit-source-id: 902cd6aff95f4b92871e87b072cce5267e7840d5