call _PyClassLoader_AddSubclass from add_subclass
Summary:
This is required in order to get vtables initialized for subclasses
of static classes. The parent class gets its vtable lazily initialized
by `classloader_init_slot`, but we may need to call the method in
that same slot on a subclass without ever going through
`classloader_init_slot` (the slot is already initialized!) so we need to
make sure subclasses always have an initialized vtable.

Reviewed By: tekknolagi

Differential Revision: D38637929

fbshipit-source-id: e46bf68b403c1bd3bada8159c52e01bb17c8b123