ensure compiler empty-block removal works with SP jumps
Summary:
In some places the base compiler (following upstream) does stuff
conditionally based on specific opcode names. Where we've
added new SP opcodes with similar behavior, this code will need
to be made more generic or extended in a subclass.

In this case, we were failing to bump a jump target past an empty
basic block because the jump was a `JUMP_IF_ZERO_OR_POP`
(or another opcode in that family) which are added in SP. Resolve
this by generically checking if `instr.is_jump()` instead.

Also needed to fix `is_jump()` implementation, which used the
passed-in `opcode` module in one place but failed to use it on the
next line.

Reviewed By: tekknolagi

Differential Revision: D38639878

fbshipit-source-id: 046dfc680be126a60302bca4e47e2326530484aa