mark types as TPFLAGS_CPYTHON_ALLOCATED
Summary:
In the classloader, before checking the primitive type code of a
Python type object (which lives in the `CINDER_EXTRA` memory)
we first make sure it is `TPFLAGS_CPYTHON_ALLOCATED`,
otherwise we just fall back to `TYPED_OBJECT`. Since we weren't
setting `TPFLAGS_CPYTHON_ALLOCATED` on any new types,
we were always falling back to `TYPED_OBJECT` instead of
getting the actual type code, and this broke `CHECK_ARGS`
checking of primitive type arguments.

Set `TPFLAGS_CPYTHON_ALLOCATED` on new types to fix this.

I am not actually sure why we need
`TPFLAGS_CPYTHON_ALLOCATED` instead of just relying on the
existing `TPFLAGS_HEAPTYPE`. The only theory so far is that we
are being extra safe in case some extension allocates a heap type
without going through Python's `type_new`? But let's leave it for
now to minimize differences with 3.8.

Reviewed By: sinancepel

Differential Revision: D38664333

fbshipit-source-id: 4b8d7825ef1d858524f51e084f9674e7702ab1ba