introduce the possibility of preload failing to the compilation pipeline
Summary: With lazy imports warmup, there is the possibility of a preload not being able to resolve a target descr (due to an import silently failing and the internal import_module call returning an exception, resulting in classloader_get_member returning null). In this case, we should not explode the jit, but instead log and refuse to compile the afflicted function.

Reviewed By: mpage

Differential Revision: D38559791

fbshipit-source-id: 5ad97570ed577c74dd37136a5266945c5834c635