Add deopt support to MATCH_KEYS implementation
Summary:
We should not attempt to handle exceptions in native code with
a `Guard` on a potentially nullable pointer. This will a) lead to a
GuardFailure instead of an exception having been raised, and b) does not
automatically refine the type of the now non-nullable pointer.

mpage notes that a more minimal implementation would replace Guard with
CheckExc, but I think that this bigger opcode change will help long-term with
optimizing match statements.

This diff also fixes some bugs with reversed parameter ordering and threads
`PyThreadState*` through the C API calls.

Reviewed By: mpage

Differential Revision: D38512707

fbshipit-source-id: d4bbed0aa98425954b59f737b940d44293929f15