move main logic of strict_modules.compileall to cinder
Summary: Context: We used to have the logic of compiling a module in strict mode and writing the bytecode file live in other repositories. The way the integration is done by patching `py_compile.compile`. Instead of asking other repos to have the heavy lifting logic, let's move this into the loader so that we can update the compileall logic as the strict loader changes.

Reviewed By: carljm

Differential Revision: D38032312

fbshipit-source-id: 7d49601598c9970eb8afa0f5e9014ab57390aebb