CheckedDict support
Summary:
This is a bit of an upstream challenge - we do some trickery with GC'd objects in the documentation, and replace a few dict checks with dict-or-checked-dict checks in dictobject.c.

I think the most likely path for upstreaming will be to treat checked dicts as a completely separate data structure in its own module long term, but that's a separate undertaking and I don't think we should block the port on it.

While porting, I took some care to rename all the checked dict functions as `Ci_` instead of `_Py`. Hopefully this makes the parts where we hook into the internals clear.

Reviewed By: czardoz

Differential Revision: D38744102

fbshipit-source-id: 5fba4881869cda7642347370048401c0322468e2