Deopt all JIT generators at JIT finalization
Summary:
Increased test coverage in D38719260 exposed another issue with
shutdown timing: JIT generators can live past `_PyJIT_Finalize()`, and if a GC
is triggered, `GenYieldPoint::visitRefs()` could access freed metadata that was
owned by the JIT. We could also run into issues if an object destructor
attempted to resume a JIT generator after `_PyJIT_Finalize()`, although I
haven't seen this happen yet.

To deal with this, I used `_PyGC_VisitObjects()` to deopt all live JIT
generators in `_PyJIT_Finalize()`. This in turn required replacing
`GenYieldPoint::pyobj_offs_` with a full `DeoptMetadata`, to allow deopting
suspended generators. As part of this, I added `DeoptReason::kYieldFrom` and a
`bool for_gen_resume` flag to `reifyFrame()`, which lets us use the same
`DeoptMetadata` for deopting and raising exceptions at `YieldFrom` (the
instruction pointer is off by one between the two cases; see `reifyFrame()` for
details). It's a little more complicated but should be worth the memory
savings.

I also changed how we compute the spill size for `InitialYield` in the register
allocator: rather than assuming anything about `InitialYield`'s inputs, simply
look at the live intervals when we encounter an `InitialYield` in the final
rewrite phase. This isn't strictly necessary with the way the rest of this diff
ended up going, but I still think it's cleaner so I left it in.

Reviewed By: jbower-fb

Differential Revision: D38669604

fbshipit-source-id: 7eaa369d4d176c46eee007d9b95de29f4fba4073