Make GuardIs a passthrough instruction
Summary:
Ran into an issue in which `GuardTypeRemoval` wasn't treating `GuardIs`
instructions correctly since they weren't considered passthrough
instructions. Upon examination, only `RefcountInsertion` actually
requires `GuardIs` to be treated as not a passthrough, so we should move
the special-casing of `GuardIs` there instead of special-casing it in
all other uses.

Reviewed By: swtaarrs

Differential Revision: D37653975

fbshipit-source-id: 3e37394ada509ae5b0b26748c3f94dfdad8e4d3e