Remove primitive enums
Summary:
Porting D37823776 to 3.10

>
This feature introduces a lot of complexity for little upside.
It is the only case where the HIR type does not always correspond with the unboxed object type,
which leads to a lot of special casing.
It requires special function entry points in the interpreter
and special boxing and error handling in the JIT,
which has led to a number of difficult-to-track bugs.
Despite all of this special attention, the feature is too cumbersome for widespread adoption,
as primitives cannot be used as list/set elements or dict keys/values,
which is a very important use case for enumeration values.
The main perf benefit of avoiding the LOAD_ATTR on the enum class
could instead be achieved through slotification of enum type objects,
so remove the feature and its complexity, overhead, and need for maintenance.
[William Meehan]

Reviewed By: carljm

Differential Revision: D38989732

fbshipit-source-id: f80fad79f6834bbb4658bdb8d489b34a59a244ca