pop None return from property fset invoke
Summary:
Our support for invoking `fset` on a property was missing a `POP_TOP` of the value
returned from the setter, which is unused. This always meant that we'd end up with
an extra unused value on the stack, but this becomes visible in 3.10 which has a new
assertion in `RETURN_VALUE` that the stack is empty.

Reviewed By: sinancepel

Differential Revision: D39045305

fbshipit-source-id: 0f8f8ce9748e4c8a84f0b0b3f3aa95a9ee50297a