Faster + more reliable ASAN testing
Summary:
As ASAN builds are so much slower they can increase the flakiness of tests which are time-sensitive. This can cause already slow ASAN tests to take even longer as flaky tests are re-run. In ASAN runs we are primarily interested in memory issues, so re-running flaky tests or even looking at actual test results at all is not of much value.

So, here we disregard any test results from running the CPython test suite. Instead, we send ASAN logs to files and fail only if anything shows up in these. This also solves another issue which is ASAN failures occurring on shutdown of test workers may not be logged as any test failure and so be missed. With this setup all ASAN errors will be caught and fail the Sandcastle job.

Reviewed By: alexmalyshev

Differential Revision: D38965228

fbshipit-source-id: 10e2264638f120e2bb346cad0e4c5aa6364de98f