pop module after importing chkdict
Summary:
`IMPORT_FROM` does not pop the module from the stack (so that you can do
several `IMPORT_FROM` in a row after a single `IMPORT_NAME`). So if you
`IMPORT_NAME` and then do some `IMPORT_FROM` you have to conclude with a
`POP_TOP`. Our code for importing `_static.chkdict` when the `checked_dicts`
compiler flag is used didn't do the `POP_TOP`. This was revealed by 3.10's new
assertion in `RETURN_VALUE` that the stack is empty.

Reviewed By: tekknolagi

Differential Revision: D39057906

fbshipit-source-id: dc704220f76da7bdc96ece9dd782180664abb0f0