respect checked_lists compiler flag in list() call
Summary:
For the `checked_dicts` compiler flag, we respected it both for dict literals and for
explicit `dict()` calls (by importing `chkdict` and shadowing the `dict` builtin with it.)
For `checked_lists` we were only handling literals, we didn't have the code in place
to shadow the `list` built-in. Add this so they behave consistently.

Reviewed By: czardoz

Differential Revision: D39059925

fbshipit-source-id: 29519d52c39900e4738154dfcb670a75c20dc3ce