"Port" after-fork hook to copy jit perfmap
Summary:
In 3.8, we hardcoded a call to `_PyJIT_AfterFork_Child()` in
`PyOS_AfterFork_Child()`. For 3.10, I opted to hook into the existing
`os.register_at_fork()` callback system instead, to avoid changing upstream
files. It's clearly designed to be used by managed code so accessing it from C
is a bit clunky and inefficient, but it happens once per process so I think
overall this is a better way to handle it.

Reviewed By: jbower-fb

Differential Revision: D39111672

fbshipit-source-id: c9609e4c8f13f5d4118bbdb621fb3d54a1cd7eda